checkmultisig command
These two commands don't have a good API.

Ideally `enrollmultisig` would
take 2 arguments: `wallet_name` and `descriptor`. It would be even
better if we had a convention to include the wallet name in a
descriptor-type thing that describes the totality of a multisig wallet.

Likewise, `checkmultisig` should just take a `descriptor` argument and
pull out all information needed by the ColdCard to check for the
presence of a matching wallet.

I choose not to implement it this way because HWI's `Descriptor` class
currently lacks the ability to parse a multisig descriptor. In order to
accept descriptors as arguments, I'd need to introduce lots of new code
to parse descriptors. Moreover, the generality provided by a descriptor
doesn't help because ColdCard is the only wallet to support these
features.